requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Project Requirements
pandas
pyaml
scikit-learn
scipy
numpy
matplotlip

